moved download logic for individual page to services
